requirements.txt
streamlit
PyPDF2
sentence-transformers
transformers
scikit-learn
faiss-cpu